moved pump to tier2 package increased energy storage changed order of checks for lava/water
