+++ src/main/java/techreborn/init/TRBlockEntities.java
@@ -50,7 +50,7 @@ import techreborn.blockentity.machine.iron.IronFurnaceBlockEntity;
 import techreborn.blockentity.machine.misc.AlarmBlockEntity;
 import techreborn.blockentity.machine.misc.ChargeOMatBlockEntity;
 import techreborn.blockentity.machine.misc.DrainBlockEntity;
-import techreborn.blockentity.machine.misc.PumpBlockEntity;
+import techreborn.blockentity.machine.tier2.PumpBlockEntity;
 import techreborn.blockentity.machine.multiblock.*;
 import techreborn.blockentity.machine.multiblock.casing.MachineCasingBlockEntity;
 import techreborn.blockentity.machine.tier0.block.BlockBreakerBlockEntity;
